refactor resolveEditUrl to material-sync
