+++ packages/tdev/material-sync/tests/assets/material.config.yaml
