+++ packages/tdev/material-sync/tests/resolveEditUrls.test.ts
@@ -1,6 +1,6 @@
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
-const TEST_MATERIAL_CONFIG_PATH = 'src/siteConfig/tests/assets/material.config.yaml';
+const TEST_MATERIAL_CONFIG_PATH = 'packages/tdev/material-sync/tests/assets/material.config.yaml';
 
 describe('resolveEditUrl', () => {
     const originalMaterialConfigPath = process.env.MATERIAL_CONFIG_PATH;
@@ -20,7 +20,7 @@ describe('resolveEditUrl', () => {
     });
 
     it('resolves edit URL for nesting level 1 for a given doc path version', async () => {
-        const { resolveEditUrl } = await import('../helpers');
+        const { resolveEditUrl } = await import('../src/helpers/resolveEditUrl');
         const getEditUrl = resolveEditUrl();
 
         const editUrl = getEditUrl({
@@ -32,7 +32,7 @@ describe('resolveEditUrl', () => {
         expect(editUrl).toBe('/docs/OF-BYOD-Basics/01-intro.mdx');
     });
     it('resolves edit URL for nesting level 2 for a given doc path version', async () => {
-        const { resolveEditUrl } = await import('../helpers');
+        const { resolveEditUrl } = await import('../src/helpers/resolveEditUrl');
         const getEditUrl = resolveEditUrl();
 
         const editUrl = getEditUrl({
@@ -44,7 +44,7 @@ describe('resolveEditUrl', () => {
         expect(editUrl).toBe('/docs/OF-BYOD-Basics/02-software/01-intro.mdx');
     });
     it('returns default edit URL for unknown versions', async () => {
-        const { resolveEditUrl } = await import('../helpers');
+        const { resolveEditUrl } = await import('../src/helpers/resolveEditUrl');
         const getEditUrl = resolveEditUrl();
 
         const editUrl = getEditUrl({
@@ -56,7 +56,7 @@ describe('resolveEditUrl', () => {
         expect(editUrl).toBe('/versioned_docs/version-28Gz/01-BYOD/02-software/01-intro.mdx');
     });
     it('returns default edit URL for unknown entry', async () => {
-        const { resolveEditUrl } = await import('../helpers');
+        const { resolveEditUrl } = await import('../src/helpers/resolveEditUrl');
         const getEditUrl = resolveEditUrl();
 
         const editUrl = getEditUrl({
@@ -68,7 +68,7 @@ describe('resolveEditUrl', () => {
         expect(editUrl).toBe('/versioned_docs/version-28Ga/03-FOO/BAR/01-intro.mdx');
     });
     it('resolves edit URL for different version', async () => {
-        const { resolveEditUrl } = await import('../helpers');
+        const { resolveEditUrl } = await import('../src/helpers/resolveEditUrl');
         const getEditUrl = resolveEditUrl();
 
         const editUrl = getEditUrl({
@@ -80,7 +80,7 @@ describe('resolveEditUrl', () => {
         expect(editUrl).toBe('/docs/WMS-WINF/index.mdx');
     });
     it('resolves nested docs correctly when having ignored entries', async () => {
-        const { resolveEditUrl } = await import('../helpers');
+        const { resolveEditUrl } = await import('../src/helpers/resolveEditUrl');
         const getEditUrl = resolveEditUrl();
 
         const resolvedEditUrl = getEditUrl({
