+++ packages/tdev/material-sync/tsconfig.json
@@ -11,6 +11,6 @@
         "resolveJsonModule": true,
         "moduleResolution": "NodeNext"
     },
-    "include": ["src/*.ts"],
+    "include": ["src/*.ts", "src/helpers/resolveEditUrl.ts", "src/helpers/index.ts"],
     "exclude": ["node_modules", "dist"]
 }
